PatchFusion/condaenv.wl4rv8xp.requirements.txt
huggingface-hub==0.20.1
timm==0.9.2
tqdm==4.66.1
wandb==0.16.1
opencv-python==4.8.1.78
imageio==2.33.1
mmengine==0.10.2
scipy==1.10.1
prettytable==3.10.0
kornia==0.7.2
transformers==4.36.2
torchmetrics==1.3.1
pandas==2.0.3